HDX-3000 triggering geopreview in package_create() if there are resources
+ refactoring of the geopreview triggering
+ refactoring actions in hdx_package
